Add confusion matrix example to tutorial notebook
