+++ cyclops/report/plot/classification.py
@@ -972,3 +972,46 @@ class ClassificationPlotter(Plotter):
         if layout is not None:
             fig.update_layout(layout)
         return fig
+
+    def plot_confusion_matrix(
+        self,
+        confusion_matrix: np.typing.NDArray[Any],
+        class_names: List[str],
+    ) -> go.Figure:
+        """Plot confusion matrix.
+
+        Parameters
+        ----------
+        confusion_matrix : np.typing.NDArray[Any]
+            confusion matrix
+        class_names : list
+            data class names
+
+        Returns
+        -------
+        go.Figure
+            plot figure
+
+        """
+        confusion_matrix = (
+            confusion_matrix.astype("float")
+            / confusion_matrix.sum(axis=1)[:, np.newaxis]
+        )
+        layout = {
+            "title": "Confusion Matrix",
+            "xaxis": {"title": "Predicted value"},
+            "yaxis": {"title": "Groundtruth value"},
+        }
+        fig = go.Figure(
+            data=go.Heatmap(
+                z=confusion_matrix,
+                x=class_names,
+                y=class_names,
+                hoverongaps=False,
+                colorscale="Greens",
+            ),
+            layout=layout,
+        )
+        fig.update_layout(height=512, width=1024)
+
+        return fig
